finished all

+++ Ass6/src/game/elements/objects/ScoreIndicator.java
@@ -13,6 +13,9 @@ import java.awt.Color;
  * The class Score indicator.
  */
 public class ScoreIndicator implements Sprite {
+    /**
+     * The constant FONT_SIZE.
+     */
     public static final int FONT_SIZE = 16;
     /**
      * The Boundary.
@@ -26,15 +29,25 @@ public class ScoreIndicator implements Sprite {
      * The Score.
      */
     private final Counter score;
+    /**
+     * The Text color.
+     */
     private final Color textColor;
+    /**
+     * The Lives.
+     */
     private final Counter lives;
+    /**
+     * The Name.
+     */
     private final String name;
 
     /**
      * Instantiates a new Score indicator.
      *
      * @param boundary the boundary
-     * @param score    the score
+     * @param name     the name
+     * @param counters the counters
      */
     public ScoreIndicator(final Rectangle boundary, final String name, final Counter... counters) {
         this.boundary = boundary;
@@ -49,7 +62,7 @@ public class ScoreIndicator implements Sprite {
     public void drawOn(final DrawSurface canvas) {
         int scoreX = (int) this.boundary.getCenter().getX();
         int livesX = (int) this.boundary.getCenter().getX() / 4;
-        int nameX = (int) ((3 * this.boundary.right()) / 4);
+        int nameX = (3 * this.boundary.right()) / 4;
         int textY = (int) this.boundary.getCenter().getY() + this.boundary.bottom() / 2 - 2;
         canvas.setColor(color);
         canvas.fillRectangle(boundary.left(), boundary.top(), boundary.getWidth(), boundary.getHeight());
